chore: split out Frontend into pretty and plain
